backend/requirements.txt
attrs==25.4.0
blinker==1.9.0
click==8.3.0
Flask==3.1.2
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
MarkupSafe==3.0.3
python-dotenv==1.2.1
referencing==0.37.0
rpds-py==0.28.0
Werkzeug==3.1.3
pytest>=8.0
psycopg[binary]==3.2.3
SQLAlchemy==2.0.36
alembic==1.14.0
bcrypt==4.2.0
flask-cors==6.0.1
